+++ kroxylicious-kubernetes/kroxylicious-operator/src/main/java/io/kroxylicious/kubernetes/operator/reconciler/virtualkafkacluster/VirtualKafkaClusterReconciler.java
@@ -399,7 +399,7 @@ public final class VirtualKafkaClusterReconciler implements
                 VirtualKafkaCluster.class)
                 .withName(PROXY_EVENT_SOURCE_NAME)
                 .withPrimaryToSecondaryMapper((VirtualKafkaCluster cluster) -> ResourcesUtil.localRefAsResourceId(cluster, cluster.getSpec().getProxyRef()))
-                .withSecondaryToPrimaryMapper(proxy -> ResourcesUtil.findReferringPrimaries(context,
+                .withSecondaryToPrimaryMapper(proxy -> ResourcesUtil.findKnownPrimariesOf(context,
                         proxy,
                         cluster -> Optional.of(cluster.getSpec().getProxyRef())))
                 .build();
@@ -410,7 +410,7 @@ public final class VirtualKafkaClusterReconciler implements
                 PROXY_CONFIG_STATE_SOURCE_NAME,
                 sharedConfigMapInformer,
                 VirtualKafkaClusterReconciler::toConfigStateResourceName,
-                configMap -> ResourcesUtil.findReferringPrimaries(context,
+                configMap -> ResourcesUtil.findKnownPrimariesOf(context,
                         configMap,
                         cluster -> Optional.of(new AnyLocalRefBuilder().withGroup("").withKind("ConfigMap")
                                 .withName(cluster.getSpec().getProxyRef().getName() + CONFIG_STATE_CONFIG_MAP_SUFFIX)